+++ src/vendors/helpers/Layers.js
@@ -37,6 +37,7 @@ export function createLayer(layerConfig, map) {
         extra: layerConfig.extra,
         order: layerConfig.order,
         splitscreen: layerConfig.splitscreen,
+        displayInLayerSwitcher: layerConfig.userVisible,
         source: new VectorTileSource({
           attributions: layerConfig.attributions,
           format: new MVT(),
@@ -56,6 +57,7 @@ export function createLayer(layerConfig, map) {
         extra: layerConfig.extra,
         order: layerConfig.order,
         splitscreen: layerConfig.splitscreen,
+        displayInLayerSwitcher: layerConfig.userVisible,
         source: new TileWMS({
           url: layerConfig.source.url,
           params: layerConfig.source.params,
@@ -76,6 +78,7 @@ export function createLayer(layerConfig, map) {
         extra: layerConfig.extra,
         order: layerConfig.order,
         splitscreen: layerConfig.splitscreen,
+        displayInLayerSwitcher: layerConfig.userVisible,
         source: new VectorSource({
           format: new GeoJSON(),
           url: layerConfig.source.url,
@@ -94,6 +97,7 @@ export function createLayer(layerConfig, map) {
         extra: layerConfig.extra,
         order: layerConfig.order,
         splitscreen: layerConfig.splitscreen,
+        displayInLayerSwitcher: layerConfig.userVisible,
         source: new XYZ({
           url: layerConfig.source.url,
         }),
@@ -111,6 +115,7 @@ export function createLayer(layerConfig, map) {
         extra: layerConfig.extra,
         order: layerConfig.order,
         splitscreen: layerConfig.splitscreen,
+        displayInLayerSwitcher: layerConfig.userVisible,
         source: new VectorSource({
           features: new GeoJSON().readFeatures(layerConfig.source.data, {
             // Ensure the features are read with the correct projection
@@ -135,6 +140,7 @@ export function createLayer(layerConfig, map) {
         extra: layerConfig.extra,
         order: layerConfig.order,
         splitscreen: layerConfig.splitscreen,
+        displayInLayerSwitcher: layerConfig.userVisible,
         source: new GeoTIFF({
           sources: [
             {
@@ -166,6 +172,7 @@ export function createLayer(layerConfig, map) {
         extra: layerConfig.extra,
         order: layerConfig.order,
         splitscreen: layerConfig.splitscreen,
+        displayInLayerSwitcher: layerConfig.userVisible,
         source: new VectorTileSource({
           projection: layerConfig.source?.projection || 'EPSG:3857',
         }),
@@ -188,6 +195,7 @@ export function createLayer(layerConfig, map) {
         extra: layerConfig.extra,
         order: layerConfig.order,
         splitscreen: layerConfig.splitscreen,
+        displayInLayerSwitcher: layerConfig.userVisible,
         source: new TileArcGISRest({
           url: layerConfig.source?.url,
           params: layerConfig.source.params || {},
@@ -207,6 +215,7 @@ export function createLayer(layerConfig, map) {
         extra: layerConfig.extra,
         order: layerConfig.order,
         splitscreen: layerConfig.splitscreen,
+        displayInLayerSwitcher: layerConfig.userVisible,
         source: new ImageArcGISRest({
           url: layerConfig.source?.url,
           ratio: layerConfig.source.ratio || 1,